Improvement and bug fix
Fixed bug where a shape would overlap another one. This was due to
checking collision before actually going to move that block. Fixed by
moving the collision detection into the move() function instead of
checking beforehand in the game loop.

Fixed flush_loop() so it doesn’t need to be executed every 1ms, using a
STOP flag that doesn’t load any actions in the queue when an action is
being executed.